refactor: JwtUtil 위치 변경
