+++ src/main/java/com/mycodingtest/authorization/util/JwtUtil.java
@@ -1,4 +1,4 @@
-package com.mycodingtest.authorization;
+package com.mycodingtest.authorization.util;
 
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.Jwts;
